% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMie_detector_integration.R
\name{calculate_detector_response}
\alias{calculate_detector_response}
\title{Calculate Detector Response}
\usage{
calculate_detector_response(
  S,
  detector = create_detector(),
  gain = 1,
  dr = 0.02,
  dphi = 10
)
}
\arguments{
\item{S}{A \code{ \link[Rscattnlay]{Scatterer}} particle description. This object
encapsulates the medium's and the particle's refractive indices, the particle's
geometry (e.g. core + shell, just core, etc), and the incident radiation wavelength.}

\item{detector}{A detector object, created using \link{create_detector}.  This
object describes the geometry of the detector as well as the polarization state
of the incident light.}

\item{gain}{A coefficient that expresses the relative conversion of photons
to scattering signal amplitude.  Default = 1.0.}

\item{dr}{The numerical integration delta_radius across the detector, whose
diameter is, by definition, 1.0.  Integration limits are 0,1.  Default = 0.02,
accurate to about 2 percent.}

\item{dphi}{The numerical integration of azimuthal angle delta_phi across the detector.
Integration limits are 0,360.  Default = 10.0, accurate to about .06 percent.}
}
\description{
This function performs the numerical integration of detected
scattering signal over the detector surface.  A significant constraint is that
the detector is assumed to be cylindrically symmetrical.

We retrieve the scattering amplitudes from Rscattnlay, and calculate the
Stokes matrix elements S11 and S12, which are required to accurately handle
polarization effects.
}
