% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie_transform.R
\name{calibrate_ssc_conversion}
\alias{calibrate_ssc_conversion}
\title{Calibrate Detector Response Curve}
\usage{
calibrate_ssc_conversion(
  detector = create_detector(),
  bead = create_SI(d = 200),
  bead_mfi = 60000,
  show = FALSE
)
}
\arguments{
\item{detector}{A description of the detector.  Default = create_detector().}

\item{bead}{A description of the calibration bead used.  Default is a silica
bead with a diameter of 200nm.}

\item{bead_mfi}{The "MFI" of the signal generated by the calibration bead.  Default = 60000.}

\item{show}{Display a graph of detector response as a function of EV diameter.  Default = FALSE.}
}
\value{
A list, with elements:
\describe{
  \item{gain}{The deduced gain factor for the detector}
  \item{lut}{A lookup table relating EV size to the corresponding detector signal}
}
}
\description{
This function uses the signal from a calibration bead to determine
the gain factor for the instrument.  This factor converts the output of
\link{calculate_detector_response} such that it reads out in the same units
as reported by the detector.

This function is a precursor to the use of the \link{mie_transform} function, which
calculates vesicle size and appends it to a \link[flowCore]{flowFrame}.
}
