% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowMie_detector_integration.R
\name{create_detector}
\alias{create_detector}
\title{Create a Model of a Flow Cytometer Scatter Detector}
\usage{
create_detector(theta_0 = 90, beta = 60, psi_0 = 90, pol = 1)
}
\arguments{
\item{theta_0}{Angle between the incident light source and the detector,in degrees.
Default = 90 (side scatter detector).}

\item{beta}{The detector acceptance half-angle, in degrees.  Default = 60.}

\item{psi_0}{Polarizaton of the incident light in the plane of incidence (PoI), in degrees.
90 degrees corresponds to perpendicular to the PoI.  0 degrees corresponds to
polarization in the PoI.  Default = 90.}

\item{pol}{Degree of (linear) polarization.  0 <= pol <= 1.0.  1.0 means fully polarized, 0.0 means
unpolarized.  Default = 1.0}
}
\value{
An object of class 'detector', with the following elements:
\describe{
  \item{beta}{The detector half-angle of acceptance}
  \item{psi_0}{The angle of polarization of incident light}
  \item{pol}{The degree of polarization of incident light}
  \item{theta_0}{Angle between the incident light source and the detector}
}
}
\description{
Create_detector creates a model of a flow cytometer scatter detector.
It is used along with a model of a scattering particle to model the detection of
light scatter due to small particles such as extracellular vesicles (EVs).
}
