% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\name{create_EV}
\alias{create_EV}
\title{Create a Model of an Extracellular Vesicle}
\usage{
create_EV(
  medium = 1.34,
  lambda = 488,
  n_core = 1.38,
  n_membrane = 1.46,
  thickness_membrane = 5,
  d
)
}
\arguments{
\item{medium}{The refractive index (RI) of the material surrounding the particle (Default = 1.34,
approximating the RI of normal saline solution)}

\item{lambda}{The wavelength of the incident light, in nm.  Default = 488.}

\item{n_core}{The refractive index of the EV lumen.  Default = 1.38.}

\item{n_membrane}{The refractive index of the EV membrane.  Default = 1.46.}

\item{thickness_membrane}{The thickness of the EV membrane, in nm.  Default = 5.0.}

\item{d}{The overall diameter (lumen + membrane) of the EV, in nm.}
}
\value{
An object of class \code{ \link[Rscattnlay]{Scatterer}}.
}
\description{
Create_EV creates a model of a 2-layer particle, suitable to describe an EV.
This is really just a convenience wrapper for \link{create_particle}, allowing the description of
a particle with total diameter `d`, surrounded by a membrane specified by its thickness.
The refractive indices of the core and membrane can be specified separately.
**Please note that this function uses the diameter, NOT the radius of the particle!**
}
