% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterer.R
\name{create_SI}
\alias{create_SI}
\title{Create a Model of Silica Bead}
\usage{
create_SI(medium = 1.34, lambda = 488, n = 1.463, d)
}
\arguments{
\item{medium}{The refractive index (RI) of the material surrounding the particle (Default = 1.34,
approximating the RI of normal saline solution)}

\item{lambda}{The wavelength of the incident light, in nm.  Default = 488.}

\item{n}{The refractive index of Silica.  Default = 1.463.}

\item{d}{The  diameter of the silica bead, in nm.}
}
\value{
An object of class \code{ \link[Rscattnlay]{Scatterer}}.
}
\description{
Create_SI creates a model of a silica bead.
This is really just a convenience wrapper for \link{create_particle} with default
refractive index corresponding to silica  The user only needs to
specify the size of the bead.   **Please note
that this function uses the diameter, NOT the radius of the particle!**
}
