% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mie_transform.R
\name{mie_transform}
\alias{mie_transform}
\title{Compute the Mie Transform}
\usage{
mie_transform(
  ff,
  ssc_param = "SSC-A",
  transformation = c("none", "biexp"),
  detector
)
}
\arguments{
\item{ff}{A \link[flowCore]{flowFrame} containing EV data.}

\item{ssc_param}{The name of the scattering parameter to use for sizing.
**NOTE: should be the same parameter used to detect the calibration bead in \link{calibrate_ssc_conversion()}.**}

\item{transformation}{What transformation was used on the scattering parameter.  Ones
currently supported include c("none", "biexp").}

\item{detector}{A detector, having been calibrated by \link{calibrate_ssc_conversion()}.}
}
\value{
A flowFrame, with a new parameter called "Size (nm)".
}
\description{
Calculate the Mie Transform, which converts a scattering signal to
a corresponding EV diameter by means of calculated scattering from a calibration particle.
}
\details{
There are several limitations of this function having to do with whether
the scattering data have been transformed or not.  Since the Mie Transform operates in linear
space, any non-linear transformation must be inverted prior to calculating the Mie Transform.
Currently, the only supported non-linear transformation is link[wadeTools]{biexp}.
}
