% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detector.R
\name{eta_uniform}
\alias{eta_uniform}
\alias{eta_van_der_pol}
\title{Model Detector Radial Efficiency}
\usage{
eta_uniform(alpha_prime, alpha)

eta_van_der_pol(alpha_prime, alpha)
}
\arguments{
\item{alpha_prime}{The variable angle over the detector (zero on the axis)}

\item{alpha}{The maximum acceptance angle of the detector}
}
\description{
\code{eta_uniform} models a uniform detector

\code{eta_van_der_pol} models a detector whose efficiency drops
off sinusoidally from 1.0 on the axis to 0.0 at the maximum acceptance angle
}
